Allow custom wasm malloc implementation
